- use typescript - share types between frontend and backend   This means moving types to the frontend or into a separate module
