+++ backend/tsconfig.json
@@ -18,7 +18,13 @@
     "typeRoots": [
       "./types",
       "node_modules/@types"
-    ]
+    ],
+    "paths": {
+      "*": [
+        "*",
+        "../client/src/*"
+      ]
+    }
   },
   "files": [
     "controller/google_controller.ts"
@@ -30,5 +36,6 @@
   ],
   "exclude": [
     "./src/public/"
-  ]
+  ],
+  
 }
